Use CoordinatorLayout for better SnackBar behavior
Anchoring the SnackBar to the CoordinatorLayout gives us:
- swipe to dismiss
- drawer being drawn on top of the SnackBar
